refactor: Referee Class 수정
- 정적 메서드 방식이 아닌 생성자 방식으로 변경 후 값을 초기화하면서 객체 생성하도록 변경
- reset() 함수를 통해 객체 초기화하도록 변경
- Strike 개수와 Ball 개수를 세는 함수 분리
